feat: 原主页移至 diagnostics 页面；原 info 页面移至主页
